# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KERN behavioral intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type DeveloperRole = (typeof DeveloperRole)[keyof typeof DeveloperRole];

export const DeveloperRole = {
  DEVELOPER: "DEVELOPER",
  TEAM_LEAD: "TEAM_LEAD",
  ADMIN: "ADMIN",
} as const;

export interface Developer {
  id: number;
  name: string;
  email: string;
  avatarUrl?: string | null;
  githubHandle?: string | null;
  teamId?: number | null;
  teamName?: string | null;
  role: DeveloperRole;
  timezone: string;
  createdAt: string;
}

export type DashboardSummaryTopActivityToday =
  | (typeof DashboardSummaryTopActivityToday)[keyof typeof DashboardSummaryTopActivityToday]
  | null;

export const DashboardSummaryTopActivityToday = {
  CODING: "CODING",
  DEBUGGING: "DEBUGGING",
  TESTING: "TESTING",
  DEVOPS: "DEVOPS",
  RESEARCHING: "RESEARCHING",
  IDLE: "IDLE",
} as const;

export type WorkSessionActivityType =
  (typeof WorkSessionActivityType)[keyof typeof WorkSessionActivityType];

export const WorkSessionActivityType = {
  CODING: "CODING",
  DEBUGGING: "DEBUGGING",
  TESTING: "TESTING",
  DEVOPS: "DEVOPS",
  RESEARCHING: "RESEARCHING",
  IDLE: "IDLE",
} as const;

export interface WorkSession {
  id: number;
  developerId: number;
  activityType: WorkSessionActivityType;
  inferredTask?: string | null;
  project?: string | null;
  language?: string | null;
  durationMinutes: number;
  commandCount: number;
  startedAt: string;
  endedAt: string;
  /** Classification confidence 0-1 */
  confidence: number;
}

export interface DashboardSummary {
  todaySessionCount: number;
  todayActiveMinutes: number;
  todayCommandCount: number;
  weeklyActiveMinutes: number;
  /** Days with at least one session */
  currentStreak: number;
  topActivityToday?: DashboardSummaryTopActivityToday;
  recentSessions: WorkSession[];
  standupReady: boolean;
}

export type ActivityTypeStatActivityType =
  (typeof ActivityTypeStatActivityType)[keyof typeof ActivityTypeStatActivityType];

export const ActivityTypeStatActivityType = {
  CODING: "CODING",
  DEBUGGING: "DEBUGGING",
  TESTING: "TESTING",
  DEVOPS: "DEVOPS",
  RESEARCHING: "RESEARCHING",
  IDLE: "IDLE",
} as const;

export interface ActivityTypeStat {
  activityType: ActivityTypeStatActivityType;
  totalMinutes: number;
  sessionCount: number;
  percentage: number;
}

export type DailyActivityPointBreakdown = { [key: string]: number };

export interface DailyActivityPoint {
  date: string;
  totalMinutes: number;
  sessionCount: number;
  breakdown: DailyActivityPointBreakdown;
}

export interface ActivityBreakdown {
  days: number;
  breakdown: ActivityTypeStat[];
  dailySeries: DailyActivityPoint[];
}

export interface SessionList {
  sessions: WorkSession[];
  total: number;
}

export type StandupSource = (typeof StandupSource)[keyof typeof StandupSource];

export const StandupSource = {
  AI: "AI",
  MANUAL: "MANUAL",
} as const;

export interface Standup {
  id: number;
  developerId: number;
  date: string;
  content: string;
  source: StandupSource;
  postedToSlack: boolean;
  postedAt?: string | null;
  generatedAt: string;
}

export interface StandupOrNull {
  standup: Standup | null;
}

export interface StandupList {
  standups: Standup[];
  total: number;
}

export interface UpdateStandupBody {
  content: string;
}

export interface SlackPostResult {
  success: boolean;
  message: string;
}

export type TeamMemberRole =
  (typeof TeamMemberRole)[keyof typeof TeamMemberRole];

export const TeamMemberRole = {
  DEVELOPER: "DEVELOPER",
  TEAM_LEAD: "TEAM_LEAD",
  ADMIN: "ADMIN",
} as const;

export type TeamMemberCurrentActivity =
  | (typeof TeamMemberCurrentActivity)[keyof typeof TeamMemberCurrentActivity]
  | null;

export const TeamMemberCurrentActivity = {
  CODING: "CODING",
  DEBUGGING: "DEBUGGING",
  TESTING: "TESTING",
  DEVOPS: "DEVOPS",
  RESEARCHING: "RESEARCHING",
  IDLE: "IDLE",
} as const;

export interface TeamMember {
  id: number;
  name: string;
  email: string;
  avatarUrl?: string | null;
  githubHandle?: string | null;
  role: TeamMemberRole;
  todayActiveMinutes: number;
  todaySessionCount: number;
  currentActivity?: TeamMemberCurrentActivity;
  lastSeenAt?: string | null;
  streak: number;
}

export interface TeamMemberList {
  members: TeamMember[];
  total: number;
}

export interface ProjectStat {
  project: string;
  totalMinutes: number;
  contributorCount: number;
}

export interface TeamSnapshot {
  teamName: string;
  totalActiveMinutesToday: number;
  totalSessionsToday: number;
  activeMembersToday: number;
  totalMembers: number;
  activityBreakdown: ActivityTypeStat[];
  topProjects: ProjectStat[];
  cachedAt: string;
}

export type GetActivityBreakdownParams = {
  days?: number;
};

export type ListSessionsParams = {
  limit?: number;
  offset?: number;
  /**
   * Filter by date (YYYY-MM-DD)
   */
  date?: string;
};

export type ListStandupsParams = {
  limit?: number;
  offset?: number;
};